fix: use i64

+++ anvil/src/eth/api.rs
@@ -1391,9 +1391,9 @@ impl EthApi {
     /// Jump forward in time by the given amount of time, in seconds.
     ///
     /// Handler for RPC call: `evm_increaseTime`
-    pub async fn evm_increase_time(&self, seconds: U256) -> Result<i128> {
+    pub async fn evm_increase_time(&self, seconds: U256) -> Result<i64> {
         node_info!("evm_increaseTime");
-        Ok(self.backend.time().increase_time(seconds.try_into().unwrap_or(u64::MAX)))
+        Ok(self.backend.time().increase_time(seconds.try_into().unwrap_or(u64::MAX)) as i64)
     }
 
     /// Similar to `evm_increaseTime` but takes the exact timestamp that you want in the next block
